fixing wildcards for localdb solver
