+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/SolverLocalDB.java
@@ -1,7 +1,9 @@
 package com.mageddo.dnsproxyserver.server.dns.solver;
 
+import com.mageddo.dnsproxyserver.config.Config;
 import com.mageddo.dnsproxyserver.config.Config.Entry.Type;
 import com.mageddo.dnsproxyserver.config.ConfigDAO;
+import com.mageddo.dnsproxyserver.server.dns.Hostname;
 import com.mageddo.dnsproxyserver.server.dns.Messages;
 import com.mageddo.dnsproxyserver.server.dns.Wildcards;
 import lombok.AllArgsConstructor;
@@ -34,13 +36,13 @@ public class SolverLocalDB implements Solver {
     final var askedHost = Messages.findQuestionHostname(query);
     for (final var host : Wildcards.buildHostAndWildcards(askedHost)) {
       stopWatch.split();
-      final var entry = this.configDAO.findEntryForActiveEnv(host.getValue());
+      final var entry = this.findEntryTo(host);
       if (entry == null) {
         log.trace(
             "status=partialNotFound, askedHost={}, time={}",
             askedHost, stopWatch.getTime() - stopWatch.getSplitTime()
         );
-        return null;
+        continue;
       }
       log.trace(
           "status=found, type={}, askedHost={}, time={}, totalTime={}",
@@ -55,6 +57,10 @@ public class SolverLocalDB implements Solver {
 
     log.trace("status=notFound, askedHost={}, totalTime={}", askedHost, stopWatch.getTime());
     return null;
+  }
+
+  Config.Entry findEntryTo(Hostname host) {
+    return this.configDAO.findEntryForActiveEnv(host.getValue());
   }
 
 }
